feat: Card sliding with CSS class
Cards will now slide when calling the `sortCards` function if the
correct CSS classes are present. These are automatically removed upon
completion of the animation.

The CSS class `card-slide-in` will signal that the card should be
animated to slide into the deck.

The CSS classes `card-slide-top`, `card-slide-bottom`, `card-slide-right`
and `card-slide-left` specify the direction. If no direction is present,
the card will not slide.